requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.6
django-debug-toolbar==3.2.2
Django-patch==2.2.19
pika==1.2.0
psycopg2==2.9.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
